moved register and unregister api functionality to relay port
