+++ node/live.go
@@ -14,7 +14,7 @@ func Register() {
 	if err != nil {
 		util.ExitGracefully(err.Error())
 	}
-	resp, err := util.StructRPCReq("http://"+c.DisIP+":"+c.DisCPort+"/v1/register", s, util.POST)
+	resp, err := util.StructRPCReq("http://"+c.DisIP+":"+c.DisRPort+"/v1/register", s, util.POST)
 	if err != nil {
 		util.ExitGracefully(err.Error())
 	}
@@ -27,7 +27,7 @@ func UnRegister(count int) {
 	if err != nil {
 		util.ExitGracefully(err.Error())
 	}
-	if _, err := util.StructRPCReq("http://"+c.DisIP+":"+c.DisCPort+"/v1/unregister", s, util.POST); err != nil {
+	if _, err := util.StructRPCReq("http://"+c.DisIP+":"+c.DisRPort+"/v1/unregister", s, util.POST); err != nil {
 		fmt.Println("Error, unable to unregister node at Pocket Incorporated's Dispatcher, trying again!")
 		time.Sleep(2)
 		if count > 5 {
