+++ rpc/relay/routes.go
@@ -18,6 +18,11 @@ func Routes() shared.Routes {
 		shared.Route{Name: "DispatchInfo", Method: "GET", Path: "/v1/dispatch", HandlerFunc: DispatchInfo},
 		shared.Route{Name: "Relay", Method: "POST", Path: "/v1/relay/", HandlerFunc: Relay},
 		shared.Route{Name: "RelayInfo", Method: "GET", Path: "/v1/relay/", HandlerFunc: RelayInfo},
+		shared.Route{Name: "Register", Method: "POST", Path: "/v1/register", HandlerFunc: Register},
+		shared.Route{Name: "UnRegister", Method: "POST", Path: "/v1/unregister", HandlerFunc: UnRegister},
+		shared.Route{Name: "RegisterInfo", Method: "GET", Path: "/v1/register", HandlerFunc: RegisterInfo},
+		shared.Route{Name: "UnRegisterInfo", Method: "GET", Path: "/v1/unregister", HandlerFunc: UnRegisterInfo},
+		shared.Route{Name: "Flags", Method: "GET", Path: "/v1/flags", HandlerFunc: Flags},
 	}
 	return routes
 }
